requirements.txt
pytest==8.1.1
requests==2.31.0
pytest-html==4.1.1
allure-pytest==2.13.3